+++ src/main/java/org/springbook/user/service/UserLevelUpgradePolicyImpl.java
@@ -3,8 +3,8 @@ package org.springbook.user.service;
 import org.springbook.user.domain.Level;
 import org.springbook.user.domain.User;
 
-import static org.springbook.user.service.UserService.MIN_LOGCOUNT_FOR_SILVER;
-import static org.springbook.user.service.UserService.MIN_RECCOMEND_FOR_GOLD;
+import static org.springbook.user.service.UserServiceImpl.MIN_LOGCOUNT_FOR_SILVER;
+import static org.springbook.user.service.UserServiceImpl.MIN_RECCOMEND_FOR_GOLD;
 
 public class UserLevelUpgradePolicyImpl implements UserLevelUpgradePolicy{
     @Override
